Develop

+++ etc/RowTable/RowTable.cpp
@@ -1,6 +1,6 @@
 #include "RowTable.h"
 
-void row_table::SetValues(double e, int nst, int sc, int nr, double t) {
+void row_table::SetValues(float e, int nst, int sc, int nr, float t) {
 	error = e; 
 	nSteps = nst; 
 	scale = sc; 
@@ -11,11 +11,11 @@ void row_table::SetValues(double e, int nst, int sc, int nr, double t) {
 std::string row_table::CreateString() {
 	std::string res;
 
-	res.append(std::to_string(time/(double)(nSteps / scale)));
+	res.append(std::to_string(time/(float)(nSteps / scale)));
 	res.append(";");
 	res.append(std::to_string(error));
 	res.append(";");
-	res.append(std::to_string(log10(time / (double)(nSteps / scale))));
+	res.append(std::to_string(log10(time / (float)(nSteps / scale))));
 	res.append(";");
 	res.append(std::to_string(log10(error)));
 	res.append(";");
